third_party/WebKit/LayoutTests/accessibility/aria-owns.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<div id="elements" role="group" aria-owns="item1 item2">
<div>Element 1</div>
<div>Element 2</div>
</div>

<div id="item1" role="checkbox"></div>
<div id="item2" role="radio"></div>

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that aria-owns identifies the correct element");

    if (window.accessibilityController) {
        var elements = accessibilityController.accessibleElementById("elements");
        var item1 = accessibilityController.accessibleElementById("item1");
        var item2 = accessibilityController.accessibleElementById("item2");

        shouldBeTrue("elements.ariaOwnsElementAtIndex(0).isEqual(item1)");
        shouldBeTrue("elements.ariaOwnsElementAtIndex(1).isEqual(item2)");
    }

</script>

</body>
</html>
