third_party/WebKit/LayoutTests/accessibility/aria-owns-expected.txt

Element 1
Element 2
This tests that aria-owns identifies the correct element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elements.ariaOwnsElementAtIndex(0).isEqual(item1) is true
PASS elements.ariaOwnsElementAtIndex(1).isEqual(item2) is true
PASS successfullyParsed is true

TEST COMPLETE
